// src/app/interfaces/historicoTimeUsuario.ts
export interface HistoricoTimeUsuario {
    nrSeqTime: number;
    nrContatoUsuario: string;
    nomeUsuario: string;
    time_id: number;
    assinante: boolean;
    foto_perfil: string;
    nome: string;
    nome_cartola: string;
    slug: string;
    url_escudo_png: string;
    url_escudo_svg: string;
    facebook_id: number;
}