// src/app/interfaces/timeLigaCartola.ts
export interface TimeLigaCartola {
    idLiga: number;
    time_id: number;
    assinante: boolean;
    foto_perfil: string;
    nomeTime: string;
    nome_cartola: string;
    slug: string;
    url_escudo_png: string;
    url_escudo_svg: string;
    facebook_id: number;
    statusPgto: string;
    pontosTotais: number;
    qtJogadoresPontuados: number;
    inPoint: boolean;

}